+++ src/config/Host.cpp
@@ -1,4 +1,4 @@
 #include "Host.hpp"
 
 void Host::setHostname(std::string hostname)	{ _hostname = hostname; }
-void Host::setPort(uint32_t port) 			{ _port = port; }
+void Host::setPort(uint16_t port) 			{ _port = port; }
